feature: Controller, Service 알맞게 수정 및 삭제에 대한 예외 추가

+++ backend/src/main/java/com/example/backend/social/reaction/bookmark/service/BookmarkService.java
@@ -11,6 +11,7 @@ import com.example.backend.entity.MemberRepository;
 import com.example.backend.entity.PostEntity;
 import com.example.backend.entity.PostRepository;
 import com.example.backend.social.reaction.bookmark.dto.CreateBookmarkResponse;
+import com.example.backend.social.reaction.bookmark.dto.DeleteBookmarkResponse;
 import com.example.backend.social.reaction.bookmark.exception.BookmarkErrorCode;
 import com.example.backend.social.reaction.bookmark.exception.BookmarkException;
 
@@ -38,22 +39,26 @@ public class BookmarkService {
 	 * 북마크 생성 메서드
 	 * memberId와 postId를 받아 BookmarkEntity 생성
 	 *
-	 * @param memberId
-	 * @param postId
-	 * @return BookmarkResponse (DTO)
+	 * @param memberId, postId
+	 * @return CreateBookmarkResponse (DTO)
 	 */
 	@Transactional
 	public CreateBookmarkResponse createBookmark(Long memberId, Long postId) {
+		// 1. 멤버가 존재하는지 검증
 		MemberEntity member = memberRepository.findById(memberId)
 			.orElseThrow(() -> new BookmarkException(BookmarkErrorCode.MEMBER_NOT_FOUND));
+
+		// 2. 게시물이 존재하는지 검증
 		PostEntity post = postRepository.findById(postId)
 			.orElseThrow(() -> new BookmarkException(BookmarkErrorCode.POST_NOT_FOUND));
 
+		// 3. id 및 생성 날짜를 포함하기 위해 build
 		BookmarkEntity bookmark = BookmarkEntity.builder()
 				.member(member)
 				.post(post)
 				.build();
 
+		// 생성 로직
 		bookmarkRepository.save(bookmark);
 
 		return CreateBookmarkResponse.toResponse(bookmark);
@@ -61,18 +66,30 @@ public class BookmarkService {
 
 	/**
 	 * 북마크 삭제 메서드
-	 * memberId와 postId를 받아 BookmarkEntity 삭제
+	 * id, memberId, postId를 받아 BookmarkEntity 삭제
 	 *
-	 * @param memberId
-	 * @param postId
+	 * @param id, memberId, postId
+	 * @return DeleteBookmarkResponse (DTO)
 	 */
 	@Transactional
-	public CreateBookmarkResponse deleteBookmark(Long memberId, Long postId) {
-		BookmarkEntity bookmark = bookmarkRepository.findByMemberIdAndPostId(memberId, postId)
+	public DeleteBookmarkResponse deleteBookmark(Long id, Long memberId, Long postId) {
+		// 1. 북마크가 실제로 존재하는지 검증
+		BookmarkEntity bookmark = bookmarkRepository.findById(id)
 			.orElseThrow(() -> new BookmarkException(BookmarkErrorCode.BOOKMARK_NOT_FOUND));
 
+		// 2. 북마크의 멤버 ID와 요청한 멤버 ID가 동일한지 검증
+		if (!bookmark.getMember().getId().equals(memberId)) {
+			throw new BookmarkException(BookmarkErrorCode.MEMBER_MISMATCH);
+		}
+
+		// 3. 북마크의 게시물 ID와 요청한 게시물 ID가 동일한지 검증
+		if (!bookmark.getPost().getId().equals(postId)) {
+			throw new BookmarkException(BookmarkErrorCode.POST_MISMATCH);
+		}
+
+		// 삭제 로직
 		bookmarkRepository.delete(bookmark);
 
-		return CreateBookmarkResponse.toResponse(bookmark);
+		return DeleteBookmarkResponse.toResponse(bookmark);
 	}
 }
